src/main/persistence/JsonWriter.java
package persistence;

import model.IncorrectAnswers;
import model.QuestionBank;
import model.ScoreBoard;
import org.json.JSONObject;

import java.io.*;

// This class references code from this repo
// Link: https://github.students.cs.ubc.ca/CPSC210/JsonSerializationDemo
// Represents a writer that writes JSON representation of scoreboard, QuestionBank, and incorrect answers to file
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;

    // EFFECTS: constructs writer to write to destination file
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer; throws FileNotFoundException if destination file cannot
    // be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of scoreboard to file
    public void write(ScoreBoard sb) {
        JSONObject json = sb.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of user-created QuestionBank to file
    public void write(QuestionBank questionBank) {
        JSONObject json = questionBank.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of user-created IncorrectAnswersBank to file
    public void write(IncorrectAnswers incorrectAnswers) {
        JSONObject json = incorrectAnswers.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
    }
}
